Renames reducer/actions from bus to routes
