+++ src/containers/SearchHeader.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
 import SearchHeaderComponent from '../components/SearchHeader';
-import { getRoutes } from '../actions/bus';
+import { getRoutes } from '../actions/routes';
 
 class SearchHeader extends PureComponent {
   static propTypes = {
